+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API.Tests/Helpers/DatabaseModelBuilder.cs
@@ -203,6 +203,7 @@ namespace Dfe.ManageFreeSchoolProjects.API.Tests.Helpers
             result.FsgPreOpeningMilestonesMfadDraftedFaHealthCheck = true;
             result.FsgPreOpeningMilestonesMfadRegionalDirectorSignedOffFaHealthCheck = false;
             result.FsgPreOpeningMilestonesMfadMinisterSignedOffFaHealthCheck = false;
+            result.FsgPreOpeningMilestonesMfadIncludedSignedOffImpactAssessmentFaHealthCheck = false;
             result.FsgPreOpeningMilestonesMfadSavedFaHealthCheckInWorkplacesFolder = false;
 
             return result;
@@ -217,6 +218,7 @@ namespace Dfe.ManageFreeSchoolProjects.API.Tests.Helpers
             result.FsgPreOpeningMilestonesMfadDraftedFaSubmission = true;
             result.FsgPreOpeningMilestonesMfadRegionalDirectorSignedOffFaSubmission = false;
             result.FsgPreOpeningMilestonesMfadMinisterSignedOffFaSubmission = false;
+            result.FsgPreOpeningMilestonesMfadIncludedSignedOffImpactAssessmentFaSubmission = false;
             result.FsgPreOpeningMilestonesMfadSavedFaSubmissionInWorkplacesFolder = false;
 
             return result;