+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.CypressTests/cypress/e2e/tasks/pre-opening/tasklist-funding-agreement-submission.cy.ts
@@ -38,6 +38,7 @@ describe("Testing Funding agreement submission Task", () => {
             .summaryShows("Drafted a funding agreement (FA) submission").IsEmpty().HasChangeLink()
             .summaryShows("Regional Director has signed-off the FA submission (if applicable)").IsEmpty().HasChangeLink()
             .summaryShows("Minister has signed-off the FA submission (if applicable)").IsEmpty().HasChangeLink()
+            .summaryShows("Included a signed-off impact assessment in the submission").IsEmpty().HasChangeLink()
             .summaryShows("Saved the submission in Workplaces folder").IsEmpty().HasChangeLink()
             .isNotMarkedAsComplete();
 
@@ -67,6 +68,7 @@ describe("Testing Funding agreement submission Task", () => {
         fundingAgreementSubmissionEditPage
             .checkDraftedFASubmission()
             .checkMinisterSignedOffFASubmission()
+            .checkIncludedSignedOffImpactAssessment()
             .clickContinue()
 
         summaryPage
@@ -76,6 +78,7 @@ describe("Testing Funding agreement submission Task", () => {
             .summaryShows("Drafted a funding agreement (FA) submission").HasValue("Yes").HasChangeLink()
             .summaryShows("Regional Director has signed-off the FA submission (if applicable)").IsEmpty().HasChangeLink()
             .summaryShows("Minister has signed-off the FA submission (if applicable)").HasValue("Yes").HasChangeLink()
+            .summaryShows("Included a signed-off impact assessment in the submission").HasValue("Yes").HasChangeLink()
             .summaryShows("Saved the submission in Workplaces folder").IsEmpty().HasChangeLink()
             .isNotMarkedAsComplete();
 
@@ -96,6 +99,7 @@ describe("Testing Funding agreement submission Task", () => {
             .summaryShows("Drafted a funding agreement (FA) submission").HasValue("Yes").HasChangeLink()
             .summaryShows("Regional Director has signed-off the FA submission (if applicable)").HasValue("Yes").HasChangeLink()
             .summaryShows("Minister has signed-off the FA submission (if applicable)").IsEmpty().HasChangeLink()
+            .summaryShows("Included a signed-off impact assessment in the submission").HasValue("Yes").HasChangeLink()
             .summaryShows("Saved the submission in Workplaces folder").HasValue("Yes").HasChangeLink()
             .isNotMarkedAsComplete()
             .MarkAsComplete()