+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.Data/Entities/Milestones.cs
@@ -185,5 +185,9 @@ namespace Dfe.ManageFreeSchoolProjects.Data.Entities.Existing
         public bool? PFACMSentAnEmailToTheTrust { get; set; }
 
         public string PFACMWhyAMeetingWasNotHeld { get; set; }
+
+        public bool? FsgPreOpeningMilestonesMfadIncludedSignedOffImpactAssessmentFaHealthCheck { get; set; }
+
+        public bool? FsgPreOpeningMilestonesMfadIncludedSignedOffImpactAssessmentFaSubmission { get; set; }
     }
 }