+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects/Pages/Project/Create/CentralRoute/ApplicationNumber.cshtml.cs
@@ -10,7 +10,8 @@ public class ApplicationNumberViewModel(ICreateProjectCache createProjectCache, 
 {
     [BindProperty(Name = "application-number")]
     [Display(Name = "application number")]
-    [StringLength(10, ErrorMessage = "The application number must be 10 characters or less.")]
+    [StringLength(10, ErrorMessage = "The application number must be 10 characters or less")]
+    [Required(ErrorMessage = "Enter the application number")]
     public string ApplicationNumber { get; set; }
     
     public IActionResult OnGet()